CORE/game *(merge conflict between modules and system manager in Game.hpp)
