+++ engine/include/ComponentsSeeker.hpp
@@ -1,0 +1,5 @@
+#ifndef __AZURITE__INNER__COMPONENTS_SEEKER
+#define __AZURITE__INNER__COMPONENTS_SEEKER
+
+
+#endif
