+++ engine/include/ComponentsStorage.impl.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "ComponentsStorage.hpp"
-
 using namespace Azurite;
 
 template<typename... T>
