+++ engine/src/ComponentsStorage.cpp
@@ -1,5 +1,4 @@
 #include "ComponentsStorage.hpp"
-#include "ComponentsStorage.impl.hpp"
 
 using namespace Azurite;
