+++ engine/src/SystemsManager.cpp
@@ -18,13 +18,25 @@ void SystemsManager::runSystems()
         system.run();
 }
 
+void SystemsManager::createCoreSystem(std::function<void(Game &)> function)
+{
+    m_systems.emplace_back(*this, function);
+}
+
 /*
 ** SYSTEM IMPLEMENTATION
 */
 
-void SystemsManager::createCoreSystem(std::function<void(Game &)> function)
+// Core System Ctor
+SystemsManager::System::System(SystemsManager &owner, std::function<void(Game &)> function)
+: m_owner(owner), m_function(function)
 {
-    m_systems.emplace_back(*this, function);
+    m_summoner = [](Game &game, std::any function) {
+        std::function<void(Game &)> casted_function = \
+        std::any_cast<std::function<void(Game &)>>(function);
+
+        casted_function(game);
+    };
 }
 
 void SystemsManager::System::run()
